requirements.txt
python-dotenv
django 
pyrogram 
tgcrypto